+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.6.1-beta.3
+* EXpanded the types allowed for the search query to include `number` in addition to `string` 
+
+## 3.6.1-beta.2
+* Fixed a bug in the CDCB `getLactationsTestDate()` url that caused the search to fail
+
+## 3.6.1-beta.1
+* Updated valid search options for the CDCB
+
 ## 3.6.0-beta.7
 * Updated Github Token permissons to allow publishing to npm with provenance